+++ src/abstraction/index.ts
@@ -1,4 +1,5 @@
 export * from './i-field';
 export * from './i-schema';
 export * from './i-timer';
+export * from './primitive';
 export * from './type';