+++ docs/general/CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.<br/>
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).<br/><br/>
 
+## [0.8.0] - [FEATURE] Added the Primitive Type (2021-11-24)
+* added the `Primitive` type to represent a `string`, `number` or `boolean` type.
+* added the `snake()` function to convert camel or kebab case to snake case.
+
 ## [0.7.3] - [SECURITY] Secured Dependency Vulnerabilities (2021-07-03)
 * upgraded the `hosted-git-info` module to version `2.8.9`.
 * upgraded the `lodash` module to version `4.17.21`.